mirsha-vet-epigeoai/requirements.txt
streamlit
pandas
geopandas
pysal
esda
folium
supabase
psycopg2-binary
python-dotenv
matplotlib
scikit-learn
branca
ipywidgets
pytest
Flask
networkx
pyvis
biopython
plotly
scipy